Bootcamp/language-drills/basic-drills/pyproject.toml
[project]
name = "helloworld"
version = "0.1.0"
description = "Say hello with typer"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = ["typer[all]", "pyyaml"]

[project.scripts]
hello = "helloworld.sayhello:say_hello"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
